Modularize frontend into modules
