+++ src/dom.mjs
@@ -1,0 +1,25 @@
+export const termsInput = document.getElementById('terms');
+export const minLikesInput = document.getElementById('minLikes');
+export const timeFilterSelect = document.getElementById('timeFilter');
+export const sortSelect = document.getElementById('sortSelect');
+export const searchBtn = document.getElementById('searchBtn');
+export const statusDiv = document.getElementById('status');
+export const newPostsDiv = document.getElementById('newPosts');
+export const resultsDiv = document.getElementById('results');
+export const autoRefreshToggle = document.getElementById('autoRefreshToggle');
+export const refreshIntervalSelect = document.getElementById('refreshInterval');
+export const refreshStateDiv = document.getElementById('refreshState');
+export const refreshLastDiv = document.getElementById('refreshLast');
+export const refreshNextDiv = document.getElementById('refreshNext');
+export const themeSelect = document.getElementById('themeSelect');
+export const expandTermsToggle = document.getElementById('expandTermsToggle');
+export const expandSummary = document.getElementById('expandSummary');
+export const quoteForm = document.getElementById('quoteForm');
+export const postUrlInput = document.getElementById('postUrl');
+export const quoteSearchBtn = document.getElementById('quoteSearchBtn');
+export const quoteStatusDiv = document.getElementById('quoteStatus');
+export const quoteTabs = document.getElementById('quoteTabs');
+export const quoteOriginalDiv = document.getElementById('quoteOriginal');
+export const quoteCountDiv = document.getElementById('quoteCount');
+export const quoteResultsDiv = document.getElementById('quoteResults');
+export const quoteLoadMoreDiv = document.getElementById('quoteLoadMore');
