+++ src/testing.mjs
@@ -1,0 +1,15 @@
+export {
+  deduplicatePosts,
+  expandSearchTerms,
+  filterByLikes,
+  formatDuration,
+  getSearchCacheKey,
+  isValidBskyUrl,
+  normalizeTerm,
+  parseBlueskyPostUrl,
+  sortPosts,
+} from './utils.mjs';
+export { trackQuoteCursor } from './quotes-state.mjs';
+export { getCachedDid } from './cache.mjs';
+export { didCache } from './state.mjs';
+export { DID_CACHE_TTL_MS } from './constants.mjs';
